Fix major bug in filesystem size calculation. (RBL-2330)
The caculation skipped troves that were first seen byDefault=False but were
also included later on byDefault=True, resulting in a substantial discrepency
in installed trove size. The new results are not exact, but err on the side of
caution. Individual file sizes are also rounded up to the nearest 4K to
account for block size.